Correct datatype for MJD, Printing header before writing output

+++ filhead.h
@@ -9,6 +9,6 @@ void send_double(FILE *outfile, const char *name, double number);
 void send_int(FILE *outfile, const char *name, int number);
 void send_long(FILE *outfile, const char *name, long number);
 
-void filterbank_header(FILE *outfile, const char *infilename, const char *jname, int mjd, double freq, double bw, int nchan, double tsmpl, int nbit, int npol);
+void filterbank_header(FILE *outfile, const char *infilename, const char *jname, double mjd, double freq, double bw, int nchan, double tsmpl, int nbit, int npol);
 
 #endif
